pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "enphase-home-api-client"
version = "0.0.1"
authors = [
  { name="Sean Humes", email="107776081+TheSchipper@users.noreply.github.com" },
]
description = "API Client for Enphase Home systems"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.3",
]